Updated names of ContentSerializer's children
